Support unix domain sockets as an alternative to TCP/IP
On some systems, the ephemeral port range, and the TCP port range in general is
a contended resource. Using a UNIX domain socket, tied to the PID, ensures that
there is never a conflict, and never a race condition with the use of the
"guessed" port.

This change relies on customisations/fixes to requests-unixsocket to improve
the monkeypatching and to be more tolerant of 'proxies'