WIP: Initial hacking on snippets for parameter completions.
Only works with UltiSnips.

Generate a snippet from the completion response if the 'chunks' entry is
provided. Chunks (and indeed snippets) are recursive, in the sense that chunks
can be embedded in other chunks. Make each chunk a snippet placeholder. Note:
This means even things like commas and brackets are placeholders, but that is OK
- anything we insert automatically can be effectively optional).

Embed the ad-hoc snippet in 'user_data' item in the complete-items entry (this
is a Vim customisation) Use that embedded snippet to trigger when CompleteDone
is called.

Unfortunately, this only really works if CompleteDone is called due to CTRL-Y
rather than due to typing another character. And this workflow/user experience
is not great. Requires more effort, and perhaps more support from Vim.

FWIW: This change works well with argument hints, which are still provided. One
option I would like to explore is to auto-trigger the argument hint completion
request when we highlight a placeholder which is an argument. Currently you have
to type whitespace to trigger it (as we're 'typing' the comma for the user as
part of the snippet)