Close the splits after editing
This leaves them as hidden buffers, allowing use of the quickfix list
(with ctrl-w <enter>) to open the files in splits.

In order to prevent this being very slow and/or leaving strange artefacts
(like location lists), we disable autocommands when opening.

FIXME/TODO:
The drawback of this is the buffers are not set up correctly when edited.