Generate coverage for python in vimscript tests
Generate and send coverage data for the embedded YCM python code when
running vimscript tests. This uses a submodule for coveragepy rather
than a pip-installed one because it was impossible to get the virtualenv
properly activated in the python interpreter embedded within Vim.

This is not ideal, as it requires all users to sync this submodule, but
it does happen to work.