Remove hard-coded list of diag fietypes from YCM
We used to have some hard-coded list of filetypes which were used to
dtermine whih filetypes the server provided diagnotics for. This was at
some point refactored to be used as a list of filetypes which provided
diagnotics in the "FileReadyToParse" event.

Most diagnotics now are asynchronous, delivered using the long-polling
receive_messages request. Instead of hard-coding a list of filetypes, we
now remember if we ever received async diagnotics for a buffer, and
exclusively use them from that point on, with the exception of
YcmForceCompileAndDiagnostics, which still uses the "FileReadyToParse"
style of diagnostics. This simplifies the code and removes a list that
we always forget to update and can't actually rely on as being
hard-coded anyway.