Remove file data from YCM log
The YCM log prints all file data to a single line for a given request.
This makes viewing the log in Vim excrutiatingly slow and painful.

Usually the file data is not interesting for the purposes of
diagnostics, so truncate it. This also (somewhat) makes the log more
easily shared if the contenst of the files are proprietary.