Reduce calls to complete (by half)
coplete() is very expensive and appears as a massive hotspot in any
profiles. It's not clear if this is worse than <C-x><C-u> because that
time is not recorded in the profiles.

Anyway, there does not appear to be a good reason to call complete twice
per keypress. The comment seems old and i could not find a way to break
it in the way that it suggests.

But in any case, for now i'm happy with a slightly buggy edge case for
generally better performance.