Improve performance: Don't refresh signature help so often
We would like to skip the call to complete() as well (i.e. call it half
as often) but this introduces a subtle bug with autoindent which may
well be a vim bug - the call to 'complete()' causes auto indentation to
happen and means that our complete request gets the position
post-auto-indent, which is what we need.