Symbol search: allow fuzzy matching output
This is crude, but allows you to fuzzy-match whatever you see on the
screen using the built-in fuzzy matcher. This can be useful when you
have 100 types named the same thing and you want to zero in on the
specific on with a specific filetype or file path.

Why don't we use FilterAndSortCandidates? No reason, though this is less
"traffic".

How do you trigger it? you type your initial query, then a `/` followed
by the local filter. This feels natural as it's "like searching the
screen".

Example: ServerConfig/fooservercpp

THings to think about:

- Too unobvious?
- Only match filename?
- Can we do multiple filters?