Fix vim tests for hierarchies
Previously the tests were using async result checking, but this is
intended for insert mode only. In fact we were running off the end of
the test and then the check callbacks were firing.

Simplified by just making it the same but sequential, and replacding
FeedAndCheck* with direct calls to feedkeys(..., 'xt').