Support virtual text for all diags in file
This shows the floaty error for all messages. This is particularly
useful for LLVM-based engines like C++ and Rust where "hints" are added
to other parts of the buffer that related to specific messages.

WIth a lot of errors, this can be a bit messy and overwhelming, but in
general the expectation is that there aren't so many.