Format on save
Add a simple API for enabling format-on-save selectively.

To use this, add the following to something like ftplugin/rust.vim (or
some FileType autocommand or anything else of your choosing):

```
autocmd BufWritePre,FileWritePre <buffer> call youcompleteme#FormatPreFileSave()
```

It's sort of TODO because FileWritePre should use a subrange and that
required range formatting to work, but who uses this "partial write"
feature anyway?